third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/styles-4/styles-url-linkify-expected.txt

Tests that URLs are linked to and completed correctly. Bugs 51663, 53171, 62643, 72373, 79905

URLs completed:
http://example.com/
http://example.com/moo
https://secure.com/moo
https://secure.com/moo
http://example.com/moo
http://example.com/foo/zoo/moo
http://example.com/foo/boo/moo
http://example.com/moo
http://example.com/foo?a=b
http://example.com/foo?a=/b
http://example.com/?a=/b
http://example.com/foo?a=b
http://example.com/foo?a=/b
http://example.com/foo?a=b
http://example.com/cat.jpeg
http://example.com/cat.jpeg
http://example.com/foo?bar=http://otherexample.com
data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAEIAAABCAgMAAACeOuh7AAAABGdBTUEAAK/INwWK6QAAAAlQTFRF////AAAA////fu+PTwAAAAF0Uk5TAEDm2GYAAACHSURBVDjLxdLbDYAgDAVQGELn0R3oEHYf2KGdUqtE46OFRCP3oyTng1xCnWsaD5JRRtCkQ2YmkBkHRXqWJBn0j0TICbrsWVoWhRShCdcGyZCtHxMaUnVPRZ9KSbmBJdsX2vJVnwqRD0Rb4rpzgIbE/AI5NTnWAMvy5l0dXrfuLh5OCe5BmmYGXhTUxlQ5xJ8AAAAASUVORK5CYII=
javascript:alert('foo');
blob:http://example.com/f91b7b00-00-0000-0000-3b7c87055d7a
blob:http://example.com/f91b7b00-00-0000-0000-3b7c87055d7a
Link for a URI from CSS document:
devtools-link elements/styles/resources/fromcss.png
Link for a URI from iframe inline stylesheet:
elements/styles/resources/iframed.png
